move mysql sink related code to the same package
