+++ downstreamadapter/sink/pulsar_sink_test.go
@@ -16,6 +16,7 @@ package sink
 import (
 	"context"
 	"net/url"
+	"sync/atomic"
 	"testing"
 	"time"
 
@@ -85,7 +86,7 @@ func TestPulsarSinkBasicFunctionality(t *testing.T) {
 	sink, dmlProducer, ddlProducer, err := newPulsarSinkForTest(t)
 	require.NoError(t, err)
 
-	count.Store(0)
+	var count atomic.Int64
 
 	helper := commonEvent.NewEventTestHelper(t)
 	defer helper.Close()
